user pundit to authorize post owner

+++ app/controllers/posts_controller.rb
@@ -1,7 +1,6 @@
 class PostsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
   before_action :set_post, only: [:show, :edit, :update, :destroy]
-  before_action :validate_post_owner, only: [:edit, :update, :destroy]
   require 'csv'
 
   def index
@@ -47,9 +46,12 @@ class PostsController < ApplicationController
 
   def show; end
 
-  def edit; end
+  def edit
+    authorize @post, :edit?, policy_class: PostPolicy
+  end
 
   def update
+    authorize @post, :update?, policy_class: PostPolicy
     if @post.update(post_params)
       redirect_to posts_path
     else
@@ -58,6 +60,7 @@ class PostsController < ApplicationController
   end
 
   def destroy
+    authorize @post, :destroy?, policy_class: PostPolicy
     redirect_to posts_path if @post.destroy
   end
 
@@ -69,12 +72,5 @@ class PostsController < ApplicationController
 
   def post_params
     params.require(:post).permit(:title, :content, :image, category_ids: [])
-  end
-
-  def validate_post_owner
-    unless @post.user == current_user
-      flash[:notice] = 'the post not belongs to you'
-      redirect_to posts_path
-    end
   end
 end
